AIRBOTS (Delivery)/scripts/comms/inter_drone_api.py

import rospy
from mavros_msgs.msg import GlobalPositionTarget

class DroneCoordinator:
    def __init__(self):
        rospy.init_node('airbots_coordinator')
        self.scout_sub = rospy.Subscriber(
            '/razor/survivor_coords',
            GlobalPositionTarget,
            self.update_delivery_path
        )
    
    def update_delivery_path(self, msg):
        # Path replanning logic using received coordinates
        pass
